refactor: fix `getOutgoingLinksForEntity` and `getIncomingLinksForEntity` methods to handle entity IDs with draft IDs gracefully
