+++ apps/hash-frontend/src/pages/shared/discard-draft-entity-button.tsx
@@ -1,8 +1,6 @@
 import { useMutation } from "@apollo/client";
 import { AlertModal } from "@hashintel/design-system";
 import type { Entity } from "@local/hash-graph-sdk/entity";
-import type { EntityId } from "@local/hash-graph-types/entity";
-import { ENTITY_ID_DELIMITER } from "@local/hash-graph-types/entity";
 import { generateEntityLabel } from "@local/hash-isomorphic-utils/generate-entity-label";
 import type { EntityRootType, Subgraph } from "@local/hash-subgraph";
 import { extractDraftIdFromEntityId } from "@local/hash-subgraph";
@@ -23,13 +21,6 @@ import { useNotificationEntities } from "../../shared/notification-entities-cont
 import type { ButtonProps } from "../../shared/ui";
 import { Button } from "../../shared/ui";
 import { useNotificationsWithLinks } from "./notifications-with-links-context";
-
-const stripDraftIdFromEntityId = (entityId: EntityId) => {
-  return entityId
-    .split(ENTITY_ID_DELIMITER)
-    .slice(0, 2)
-    .join(ENTITY_ID_DELIMITER) as EntityId;
-};
 
 export const DiscardDraftEntityButton: FunctionComponent<
   {
@@ -97,25 +88,20 @@ export const DiscardDraftEntityButton: FunctionComponent<
 
   const isLinkEntity = !!draftEntity.linkData;
 
-  const draftEntityIdWithoutDraftId = useMemo(
-    () => stripDraftIdFromEntityId(draftEntity.metadata.recordId.entityId),
-    [draftEntity],
-  );
-
   const incomingDraftLinks = useMemo(
     () =>
       isLinkEntity
         ? undefined
         : getIncomingLinksForEntity(
             draftEntitySubgraph,
-            draftEntityIdWithoutDraftId,
+            draftEntity.metadata.recordId.entityId,
           ).filter(
             (linkEntity) =>
               !!extractDraftIdFromEntityId(
                 linkEntity.metadata.recordId.entityId,
               ),
           ),
-    [draftEntitySubgraph, draftEntityIdWithoutDraftId, isLinkEntity],
+    [draftEntitySubgraph, draftEntity, isLinkEntity],
   );
 
   const outgoingDraftLinks = useMemo(
@@ -124,14 +110,14 @@ export const DiscardDraftEntityButton: FunctionComponent<
         ? undefined
         : getOutgoingLinksForEntity(
             draftEntitySubgraph,
-            draftEntityIdWithoutDraftId,
+            draftEntity.metadata.recordId.entityId,
           ).filter(
             (linkEntity) =>
               !!extractDraftIdFromEntityId(
                 linkEntity.metadata.recordId.entityId,
               ),
           ),
-    [draftEntitySubgraph, draftEntityIdWithoutDraftId, isLinkEntity],
+    [draftEntitySubgraph, draftEntity, isLinkEntity],
   );
 
   const hasIncomingOrOutgoingDraftLinks = useMemo(
